+++ CHANGES.md
@@ -5,6 +5,8 @@ Beta 5 (in progress)
 --------------------
 * added up()
 * added not()
+* formatNumber() supports any character as group separator
+* formatNumber() will limit the number of digits in front of the decimal separator, unless you use a single '#'
 * a promise's state can now be obtained by invoking it as a function without arguments (see promise(), full package only)
 
 Beta 4.2 (2014-12-26)